Fix to pull 420

+++ AFNetworking/AFHTTPRequestOperation.m
@@ -77,11 +77,11 @@ static NSString * AFStringFromIndexSet(NSIndexSet *indexSet) {
         }
 
         if (range.length == 1) {
-            [string appendFormat:@"%lu", range.location];
+            [string appendFormat:@"%lu", (long)range.location];
         } else {
             NSUInteger firstIndex = range.location;
             NSUInteger lastIndex = firstIndex + range.length - 1;
-            [string appendFormat:@"%lu-%lu", firstIndex, lastIndex];
+            [string appendFormat:@"%lu-%lu", (long)firstIndex, (long)lastIndex];
         }
 
         range.location = nextIndex;
